backend/requirements.txt
Django==5.1.8
django-cors-headers==4.7.0
djangorestframework==3.16.0

django-vite==3.1.0

psycopg2-binary==2.9.10